context/peerbots-docs/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
idna==3.10; python_version >= '3.6'
pyaudio==0.2.14
pydantic==2.10.3
pydantic-core==2.27.1; python_version >= '3.8'
python-dotenv==1.0.1
requests==2.32.3
speechrecognition==3.12.0
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
